Proper checking for `np.array == None`
The current codebase checks `np.array == None` in a lot of places, for instance for body forces in core.planqe. This causes an exception in numpy, crashing the program. 

This change instead uses the `np.array is None` pattern to do such checks, as it instead compares the type of the array to None, rather than the value (which == does).